Tweak body of sumAll function in .js file
difficulty connecting both conditions
successfully passed the first three tests, but when
I ran all tests only the last three passed.

looked at solution to understand the use of a temp variable
and if i should put the typeof statement and check if < 0 at the
start of the function.

Overall got around 90 % of the problem right, was just having trouble
sorting out the return statements, the loop was written correctly
and I used comparison operator ( || ) in the first two conditional
statements.